Updated script to get genetic variant clin sig
An updated script that uses new functionality of the Biomedical Data Commons API to more easily retrieve the clinical significance for each genetic variant in an input list. It then counts and prints the number of genetic variants of each clinical significance type. Then writes the results to an output file.